requirements.txt
google-auth-oauthlib>=1.2.0
langchain-google-community>=2.0.0
google-auth>=2.20.0
google-api-python-client>=2.100.0